Fixed the python setup script
